# lib/api-zod/src/generated/types/logoutRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogoutRequest {
  /**
   * If omitted, every active session for the user is revoked.
   * @minLength 1
   */
  refreshToken?: string;
}