# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface UserProfile {
  id: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  merchantId: string | null;
  merchantName: string | null;
  lastLoginAt: Date | null;
}